src/standalone_cli/requirements.txt
click
sh
lxml
python-dotenv
